fix: remove pinned items request from chart and dashboard page
