+++ packages/frontend/src/pages/Dashboard.tsx
@@ -1,6 +1,5 @@
 import {
     DashboardTileTypes,
-    ResourceViewItemType,
     type DashboardTab,
     type DashboardTile,
     type Dashboard as IDashboard,
@@ -31,8 +30,6 @@ import {
 } from '../hooks/dashboard/useDashboard';
 import useDashboardStorage from '../hooks/dashboard/useDashboardStorage';
 import { useOrganization } from '../hooks/organization/useOrganization';
-import { useDashboardPinningMutation } from '../hooks/pinning/useDashboardPinningMutation';
-import { usePinnedItems } from '../hooks/pinning/usePinnedItems';
 import useToaster from '../hooks/toaster/useToaster';
 import { useSpaceSummaries } from '../hooks/useSpaces';
 import useApp from '../providers/App/useApp';
@@ -126,26 +123,6 @@ const Dashboard: FC = () => {
     const [isDuplicateModalOpen, duplicateModalHandlers] = useDisclosure();
     const [isExportDashboardModalOpen, exportDashboardModalHandlers] =
         useDisclosure();
-    const { mutate: toggleDashboardPinning } = useDashboardPinningMutation();
-    const { data: pinnedItems } = usePinnedItems(
-        projectUuid,
-        dashboard?.pinnedListUuid ?? undefined,
-    );
-
-    const handleDashboardPinning = useCallback(() => {
-        if (!dashboardUuid) return;
-        toggleDashboardPinning({ uuid: dashboardUuid });
-    }, [dashboardUuid, toggleDashboardPinning]);
-
-    const isPinned = useMemo(() => {
-        return Boolean(
-            pinnedItems?.some(
-                (item) =>
-                    item.type === ResourceViewItemType.DASHBOARD &&
-                    item.data.uuid === dashboardUuid,
-            ),
-        );
-    }, [dashboardUuid, pinnedItems]);
 
     const hasNewSemanticLayerChart = useMemo(() => {
         if (!dashboardTiles) return false;
@@ -600,7 +577,6 @@ const Dashboard: FC = () => {
                         isSaving={isSaving}
                         oldestCacheTime={oldestCacheTime}
                         isFullscreen={isFullscreen}
-                        isPinned={isPinned}
                         activeTabUuid={activeTab?.uuid}
                         dashboardTabs={dashboardTabs}
                         isFullScreenFeatureEnabled={isFullScreenFeatureEnabled}
@@ -658,7 +634,6 @@ const Dashboard: FC = () => {
                         onDelete={deleteModalHandlers.open}
                         onExport={exportDashboardModalHandlers.open}
                         setAddingTab={setAddingTab}
-                        onTogglePin={handleDashboardPinning}
                         onEditClicked={handleEnterEditMode}
                     />
                 }
