+++ packages/frontend/src/pages/SavedExplorer.tsx
@@ -1,7 +1,5 @@
+import { useEffect } from 'react';
 import { useParams } from 'react-router';
-
-import { ResourceViewItemType } from '@lightdash/common';
-import { useCallback, useEffect, useMemo } from 'react';
 import Explorer from '../components/Explorer';
 import ExplorePanel from '../components/Explorer/ExplorePanel';
 import SavedChartsHeader from '../components/Explorer/SavedChartsHeader';
@@ -9,8 +7,6 @@ import ErrorState from '../components/common/ErrorState';
 import Page from '../components/common/Page/Page';
 import SuboptimalState from '../components/common/SuboptimalState/SuboptimalState';
 import useDashboardStorage from '../hooks/dashboard/useDashboardStorage';
-import { useChartPinningMutation } from '../hooks/pinning/useChartPinningMutation';
-import { usePinnedItems } from '../hooks/pinning/usePinnedItems';
 import { useSavedQuery } from '../hooks/useSavedQuery';
 import useApp from '../providers/App/useApp';
 import ExplorerProvider from '../providers/Explorer/ExplorerProvider';
@@ -19,7 +15,7 @@ import { ExplorerSection } from '../providers/Explorer/types';
 const SavedExplorer = () => {
     const { health } = useApp();
 
-    const { savedQueryUuid, mode, projectUuid } = useParams<{
+    const { savedQueryUuid, mode } = useParams<{
         savedQueryUuid: string;
         projectUuid: string;
         mode?: string;
@@ -32,27 +28,6 @@ const SavedExplorer = () => {
     const { data, isInitialLoading, error } = useSavedQuery({
         id: savedQueryUuid,
     });
-
-    const { mutate: togglePinChart } = useChartPinningMutation();
-    const { data: pinnedItems } = usePinnedItems(
-        projectUuid,
-        data?.pinnedListUuid ?? undefined,
-    );
-
-    const handleChartPinning = useCallback(() => {
-        if (!savedQueryUuid) return;
-        togglePinChart({ uuid: savedQueryUuid });
-    }, [savedQueryUuid, togglePinChart]);
-
-    const isPinned = useMemo(() => {
-        return Boolean(
-            pinnedItems?.some(
-                (item) =>
-                    item.type === ResourceViewItemType.CHART &&
-                    item.data.uuid === data?.uuid,
-            ),
-        );
-    }, [data?.uuid, pinnedItems]);
 
     useEffect(() => {
         // If the saved explore is part of a dashboard, set the dashboard chart info
@@ -116,12 +91,7 @@ const SavedExplorer = () => {
         >
             <Page
                 title={data?.name}
-                header={
-                    <SavedChartsHeader
-                        onTogglePin={handleChartPinning}
-                        isPinned={isPinned}
-                    />
-                }
+                header={<SavedChartsHeader />}
                 sidebar={<ExplorePanel />}
                 isSidebarOpen={isEditMode}
                 withFullHeight
